fixed or suppressed all -Xlint warnings

+++ graal/com.oracle.graal.compiler.test/src/com/oracle/graal/compiler/test/BoxingEliminationTest.java
@@ -60,7 +60,7 @@ public class BoxingEliminationTest extends GraalCompilerTest {
     }
 
     public static Object boxedObjectInteger() {
-        return (int) 1;
+        return 1;
     }
 
     public static Integer boxedInteger() {
